feat(kernel-language-model-service): Apply type suggestions
