+++ packages/kernel-language-model-service/src/ollama/base.ts
@@ -1,12 +1,13 @@
-import type { GenerateRequest, GenerateResponse } from 'ollama';
+import type { GenerateResponse, ListResponse } from 'ollama';
 
-import type {
-  InstanceConfig,
-  LanguageModel,
-  LanguageModelService,
-} from '../types.ts';
+import type { LanguageModelService } from '../types.ts';
 import { parseModelConfig } from './parse.ts';
-import type { OllamaClient, OllamaModelOptions } from './types.ts';
+import type {
+  OllamaInstanceConfig,
+  OllamaModel,
+  OllamaClient,
+  OllamaModelOptions,
+} from './types.ts';
 
 /**
  * It is recommended to create an Ollama client per model session.
@@ -25,14 +26,12 @@ export class OllamaBaseLanguageModelService<Ollama extends OllamaClient>
     this.#makeClient = makeClient;
   }
 
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  async getModels() {
+  async getModels(): Promise<ListResponse> {
     const client = await this.#makeClient();
     return await client.list();
   }
 
-  // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-  async makeInstance(config: InstanceConfig<OllamaModelOptions>) {
+  async makeInstance(config: OllamaInstanceConfig): Promise<OllamaModel> {
     const modelInfo = parseModelConfig(config);
     const { model } = modelInfo;
     const ollama = await this.#makeClient();
@@ -49,11 +48,11 @@ export class OllamaBaseLanguageModelService<Ollama extends OllamaClient>
       getInfo: async () => modelInfo,
       load: async () => {
         // eslint-disable-next-line @typescript-eslint/naming-convention
-        await ollama.generate({ model, keep_alive: -1 } as GenerateRequest);
+        await ollama.generate({ model, keep_alive: -1, prompt: '' });
       },
       unload: async () => {
         // eslint-disable-next-line @typescript-eslint/naming-convention
-        await ollama.generate({ model, keep_alive: 0 } as GenerateRequest);
+        await ollama.generate({ model, keep_alive: 0, prompt: '' });
       },
       sample: async (prompt: string, options?: Partial<OllamaModelOptions>) => {
         const response = await ollama.generate({
@@ -69,6 +68,6 @@ export class OllamaBaseLanguageModelService<Ollama extends OllamaClient>
         })();
       },
     };
-    return instance as LanguageModel<OllamaModelOptions, GenerateResponse>;
+    return harden(instance);
   }
 }
