+++ packages/kernel-language-model-service/src/ollama/parse.test.ts
@@ -1,21 +1,20 @@
 import { describe, it, expect } from 'vitest';
 
-import type { InstanceConfig } from '../types.ts';
 import { parseModelConfig } from './parse.ts';
-import type { OllamaModelOptions } from './types.ts';
+import type { OllamaInstanceConfig } from './types.ts';
 
 describe('parseModelConfig', () => {
   it('should return expected model', () => {
-    const config = { model: 'llama2:7b' } as InstanceConfig<OllamaModelOptions>;
+    const config: OllamaInstanceConfig = { model: 'llama2:7b' };
     const modelInfo = parseModelConfig(config);
     expect(modelInfo).toMatchObject({ model: 'llama2:7b' });
   });
 
   it('should return expected model with options', () => {
-    const config = {
+    const config: OllamaInstanceConfig = {
       model: 'llama2:7b',
       options: { temperature: 0.5 },
-    } as InstanceConfig<OllamaModelOptions>;
+    };
     const modelInfo = parseModelConfig(config);
     expect(modelInfo).toMatchObject({
       model: 'llama2:7b',
@@ -24,7 +23,8 @@ describe('parseModelConfig', () => {
   });
 
   it('should throw when no model is provided', () => {
-    const config = {} as InstanceConfig<OllamaModelOptions>;
+    // @ts-expect-error - destructive test
+    const config: OllamaInstanceConfig = {};
     expect(() => parseModelConfig(config)).toThrow(/model/u);
   });
 });
